requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.2.0
psycopg2-binary==2.9.9
supabase